made java implementation a separate library
