+++ java/common/src/io/nayuki/qrcodegen/BitBuffer.java
