+++ java/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Wed Apr 05 10:32:27 CST 2017
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-3.3-bin.zip
