+++ java/javase/build.gradle
@@ -1,0 +1,3 @@
+dependencies {
+    compile project(':common')
+}
